+++ src/index.ts
@@ -4,6 +4,7 @@
 
 export { useToggle } from "./useToggle";
 export { useDebounce } from "./useDebounce";
+export { useDebounceFn } from "./useDebounceFn";
 export { useThrottle } from "./useThrottle";
 export { usePrevious } from "./usePrevious";
 export { useWindowSize } from "./useWindowSize";